# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ESD Ticket Booking Platform API
  description: |
    Microservices-based event ticket booking platform for IS213.

    All responses follow a standard envelope:
    - **Success**: `{"code": N, "data": {...}}`
    - **Error**: `{"code": N, "message": "..."}`

    All endpoints accept an optional `X-Correlation-ID` header (UUID). If not provided, one is auto-generated for distributed tracing.
  version: 1.0.0
  contact:
    name: ESD Team
servers:
  - url: http://localhost:8000
    description: Kong API Gateway (local development)
...
paths:
...
  /api/waitlist/join:
    post:
      tags: [Waitlist]
      summary: Join event waitlist
      description: Adds the user to the waitlist for a sold-out event. Returns 409 if already on waitlist.
      operationId: joinWaitlist
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinWaitlistRequest"
      responses:
        "201":
          description: Added to waitlist
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/JoinWaitlistResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          description: User already on waitlist for this event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/waitlist/position/{event_id}/{user_id}:
    get:
      tags: [Waitlist]
      summary: Get waitlist position
      description: Returns the user's current position on the waitlist, along with promotion details if applicable.
      operationId: getWaitlistPosition
      parameters:
        - $ref: "#/components/parameters/CorrelationID"
        - name: event_id
          in: path
          required: true
          schema:
            type: integer
          description: Event ID
        - name: user_id
          in: path
          required: true
          schema:
            type: string
          description: User ID (Firebase UID)
      responses:
        "200":
          description: Waitlist position
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/WaitlistPosition"
        "404":
          description: User not on waitlist for this event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "500":
          $ref: "#/components/responses/InternalError"

  # ─────────────────────────────────────────────
  # Fees (Charging Service)
  # ─────────────────────────────────────────────
...
components:
  parameters:
    CorrelationID:
      name: X-Correlation-ID
      in: header
      required: false
      schema:
        type: string
        format: uuid
      description: Correlation ID for distributed tracing. Auto-generated if not provided.
  responses:
    BadRequest:
      description: Invalid request parameters or body
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    SuccessEnvelope:
      type: object
      required: [code]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 200
    ErrorEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          description: HTTP status code
          example: 400
        message:
          type: string
          description: Human-readable error message
          example: "Invalid request parameters"

    # ── Event ───────────────────────────────────
...
    JoinWaitlistRequest:
      type: object
      required: [event_id, user_id, email]
      properties:
        event_id:
          type: integer
          example: 1
        user_id:
          type: string
          example: "firebase_uid_abc123"
        email:
          type: string
          format: email
          example: "user@example.com"
        phone:
          type: string
          example: "+6591234567"
        preferred_section:
          type: string
          example: "CAT 1"
    JoinWaitlistResponse:
      type: object
      properties:
        entry_id:
          type: integer
          example: 12
        position:
          type: integer
          example: 3
        event_id:
          type: integer
          example: 1
    WaitlistPosition:
      type: object
      properties:
        entry_id:
          type: integer
          example: 12
        position:
          type: integer
          example: 3
        status:
          type: string
          example: "waiting"
        promoted_seat_id:
          type: integer
          nullable: true
          description: Seat ID assigned upon promotion (null if not yet promoted)
          example: null
        promotion_expires_at:
          type: string
          format: date-time
          nullable: true
          description: Deadline to accept promoted seat (null if not yet promoted)
          example: null

    # ── Fees (Charging) ────────────────────────
...